requirements.txt
Flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.3.2 